current working version before cleaning

+++ src/napari_spine_tracker/refinement_utils/manager.py
@@ -93,10 +93,10 @@ class TrackletManager:
         rows_idxs = data_img[data_img['id'].isin(ids.astype(int))].index
         ymins, xmins = np.array(shapes_layer_data).min(axis=1).T
         ymaxs, xmaxs = np.array(shapes_layer_data).max(axis=1).T
-        self.data.loc[rows_idxs, 'xmin'] = xmins
-        self.data.loc[rows_idxs, 'xmax'] = xmaxs
-        self.data.loc[rows_idxs, 'ymin'] = ymins
-        self.data.loc[rows_idxs, 'ymax'] = ymaxs
+        self.data.loc[rows_idxs, 'xmin'] = xmins.astype(int)
+        self.data.loc[rows_idxs, 'xmax'] = xmaxs.astype(int)
+        self.data.loc[rows_idxs, 'ymin'] = ymins.astype(int)
+        self.data.loc[rows_idxs, 'ymax'] = ymaxs.astype(int)
 
     def get_data(self):
         return self.data
